[#780] Extract methods out of HonoClient into new HonoApplicationClient.
